+++ src/main/java/com/sesac/carematching/chat/pubsub/NotificationSubscriber.java
@@ -1,4 +1,4 @@
-package com.sesac.carematching.chat.listener;
+package com.sesac.carematching.chat.pubsub;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Component;
 @Slf4j
 @RequiredArgsConstructor
 @Component
-public class RedisSubscriber {
+public class NotificationSubscriber {
 
     private final SimpMessagingTemplate messagingTemplate;
